requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Environment and configuration
python-dotenv==1.0.0

# LLM providers
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
groq==0.4.1

# HTTP requests
requests==2.31.0

# RAG and embeddings
chromadb==0.4.18

# File processing
PyPDF2==3.0.1
openpyxl==3.1.2
pandas==2.1.4
python-docx==1.1.0

# Export functionality
reportlab==4.0.7
xlsxwriter==3.1.9

# CORS middleware
fastapi-cors==0.0.6

# Additional utilities
pydantic==2.5.0
typing-extensions==4.8.0